使用TextureView

+++ src/texture/RenderThread.h
@@ -40,7 +40,7 @@ public:
     void initialize();
     void doFree();
     Q_SLOT void mouseEvent(float x, float y, int state);
-    Q_SIGNAL void textureReady(int id, const QSize& size);
+    Q_SIGNAL void textureReady(intptr_t id, const QSize& size);
     Q_SLOT void renderNext();
     intptr_t renderTargetID;
     bool isInitialized;
